fluxo de atestado

+++ src/br/com/saude/api/model/business/ConvocacaoBo.java
@@ -180,7 +180,7 @@ public class ConvocacaoBo extends GenericBo<Convocacao, ConvocacaoFilter, Convoc
 	}
 	
 	@SuppressWarnings("resource")
-	public String processarConvocacao(Convocacao convocacao) throws Exception {
+	private String[] processarConvocacao(Convocacao convocacao, boolean gerarZip) throws Exception {
 		//OBTER O HTML DO RELATÓRIO		
 		StringBuilder html = new StringBuilder();
 		String line;
@@ -197,7 +197,7 @@ public class ConvocacaoBo extends GenericBo<Convocacao, ConvocacaoFilter, Convoc
 		bufferedReader.close();
 		
 		//CONFIGURAR DIRETÓRIO DOS PDFs
-		URI pdfUri;
+		URI pdfUri = null;
 		File pdf;
 		uri = new URI(getClass().getProtectionDomain().getCodeSource().getLocation().toString()
 				+"convocacao/"+convocacao.getId());
@@ -312,48 +312,58 @@ public class ConvocacaoBo extends GenericBo<Convocacao, ConvocacaoFilter, Convoc
 			stream.close();
 		}
 		
-		
-		//GERAR ZIP
-		file = new File(file.getPath());
-		URI zipUri = new URI(getClass().getProtectionDomain().getCodeSource().getLocation()
-									.toString()+"convocacao/"+convocacao.getTitulo().replace(' ', '_')+".zip");
-		File zipFile = new File(zipUri.getPath());
-		
-		ZipEntry zipEntry;
-		FileInputStream in;
-		FileOutputStream stream = new FileOutputStream(zipFile);
-		ZipOutputStream zipStream = new ZipOutputStream(stream);
-		byte[] buffer = new byte[1024];
-		
-		for(File f : file.listFiles() ) {
-			zipEntry = new ZipEntry( f.getName() );
-			zipStream.putNextEntry(zipEntry);
-			
-			in = new FileInputStream( f.getAbsolutePath() );
-			
-			int len;
-        	while ((len = in.read(buffer)) > 0) {
-        		zipStream.write(buffer, 0, len);
-        	}
-
-        	in.close();
-		}
-
-		zipStream.closeEntry();
-		zipStream.close();
-		FileUtils.cleanDirectory(file);
-				
-		//RETORNO
-		byte[] zipArray = new byte[(int) zipFile.length()];
-		new FileInputStream(zipFile).read(zipArray);
-		
 		//RETIRAR OS REGISTROS CUJO ID DO EXAME == 0, DEPOIS SALVA CONVOCAÇÃO
 		convocacao.getEmpregadoConvocacoes().forEach(eC -> {
 			eC.getEmpregadoConvocacaoExames().removeIf(x->x.getExame().getId() == 0);
 		});
 		this.save(convocacao);
 		
-		return Base64.getEncoder().encodeToString(zipArray);
+		if(gerarZip) {
+			//GERAR ZIP
+			file = new File(file.getPath());
+			URI zipUri = new URI(getClass().getProtectionDomain().getCodeSource().getLocation()
+										.toString()+"convocacao/"+convocacao.getTitulo().replace(' ', '_')+".zip");
+			File zipFile = new File(zipUri.getPath());
+			
+			ZipEntry zipEntry;
+			FileInputStream in;
+			FileOutputStream stream = new FileOutputStream(zipFile);
+			ZipOutputStream zipStream = new ZipOutputStream(stream);
+			byte[] buffer = new byte[1024];
+			
+			for(File f : file.listFiles() ) {
+				zipEntry = new ZipEntry( f.getName() );
+				zipStream.putNextEntry(zipEntry);
+				
+				in = new FileInputStream( f.getAbsolutePath() );
+				
+				int len;
+	        	while ((len = in.read(buffer)) > 0) {
+	        		zipStream.write(buffer, 0, len);
+	        	}
+
+	        	in.close();
+			}
+
+			zipStream.closeEntry();
+			zipStream.close();
+			FileUtils.cleanDirectory(file);
+			
+			//RETORNO
+			byte[] zipArray = new byte[(int) zipFile.length()];
+			new FileInputStream(zipFile).read(zipArray);
+			
+			return new String[] {zipUri.getPath(), Base64.getEncoder().encodeToString(zipArray)};
+		}else 
+			return new String[] {pdfUri.getPath(), ""};
+	}
+	
+	public String processarConvocacaoBase64(Convocacao convocacao) throws Exception {
+		return processarConvocacao(convocacao,true)[1];
+	}
+	
+	public String processarConvocacaoPath(Convocacao convocacao) throws Exception {
+		return processarConvocacao(convocacao,false)[0];
 	}
 	
 	@Override
